requirements.txt
# Core OpenVINO và AI
openvino>=2024.0.0
openvino-genai>=2024.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Web interface
gradio>=4.0.0

# Model management
huggingface-hub>=0.17.0
transformers>=4.35.0
diffusers>=0.24.0

# Training dependencies  
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.24.0
xformers>=0.0.22

# Data processing
datasets>=2.14.0
pandas>=2.0.0

# Evaluation metrics
lpips>=0.1.4
torchmetrics>=1.2.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ControlNet dependencies
mediapipe>=0.10.0  # For pose detection
controlnet-aux>=0.0.6  # Control preprocessors (optional)
